+++ src/components/title/Title.tsx
@@ -1,5 +1,6 @@
 import { useContext } from "react";
 import { PickerContext } from "../../store/pickerContext";
+import { TITLE } from "../../constants/classNames";
 import { classJoin } from "../../utils/classJoin";
 import { addCalendarMonths } from "../../utils/datePicker";
 import { TTitleProps } from "./types";
@@ -39,6 +40,7 @@ function Title({
   return (
     <div
       className={classJoin(
+        TITLE,
         "rhmdp-text-2xl rhmdp-font-bold rhmdp-cursor-pointer",
         className
       )}